Move list-players param parser out of the route module.
Next.js rejects non-handler exports from route.ts; keep parseListPlayersSearchParams in lib so build typechecking passes.
